+++ changelog.md
@@ -11,6 +11,7 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 * `MainchainAddress` is now serialized as plain String instead of hexstring of its bytes
 * Wizards use optional env vars: PC_CHAIN_CONFIG_FILE and PC_RESOURCES_CONFIG_FILE env variables to
 locate the chain config and resources config files
+* partner-chains-demo-node will now reject commands that build chain spec when no `--chain` is specified
 * `pallet-block-producer-metadata` is updated with a configurable fee for inserting the metadata, to make attacks on unbounded storage economically infeasible
 * Added `valid_before` argument to the signed message and all extrinsics in `pallet_block_producer_metadata`. This is to
 prevent unauthorized re-submission of metadata updates. The `sign-block-producer-metadata` command was updated to